DPP-268 PoC in memory fan-out - interfaces
