+++ ledger/participant-integration-api/src/main/scala/platform/store/completions/CompletionsDao.scala
@@ -1,9 +1,6 @@
 package com.daml.platform.store.completions
 
-import akka.NotUsed
-import akka.stream.scaladsl.Source
 import com.daml.ledger.ApplicationId
-import com.daml.ledger.api.v1.command_completion_service.CompletionStreamResponse
 import com.daml.ledger.participant.state.v1.Offset
 import com.daml.lf.data.Ref
 import com.daml.logging.LoggingContext
@@ -12,16 +9,16 @@ import com.daml.platform.store.dao.CommandCompletionsTable.CompletionStreamRespo
 import scala.concurrent.Future
 
 trait CompletionsDao {
-  def getCommandCompletions(
+  def getFilteredCompletions(
       startExclusive: Offset,
       endInclusive: Offset,
       applicationId: ApplicationId,
       parties: Set[Ref.Party],
   )(implicit
       loggingContext: LoggingContext
-  ): Source[(Offset, CompletionStreamResponse), NotUsed]
+  ): Future[List[(Offset, CompletionStreamResponseWithParties)]]
 
-  def getAllCommandCompletions(
+  def getAllCompletions(
       startExclusive: Offset,
       endInclusive: Offset,
   )(implicit
